+++ apps/hub/src/evals/data/evaluators.ts
@@ -1,4 +1,6 @@
 import { Prisma } from "@quri/hub-db";
+import { LlmConfig } from "@quri/squiggle-ai";
+import { llmConfigSchema } from "@quri/squiggle-ai/server";
 
 import { prisma } from "@/lib/server/prisma";
 import { checkRootUser } from "@/users/auth";
@@ -13,32 +15,57 @@ export const evaluatorSelect = {
   config: true,
   _count: {
     select: {
-      Eval: true,
+      evals: true,
     },
   },
 } satisfies Prisma.EvaluatorSelect;
 
 // Type for evaluator data
-export type Evaluator = Prisma.EvaluatorGetPayload<{
+type DbEvaluator = Prisma.EvaluatorGetPayload<{
   select: typeof evaluatorSelect;
 }>;
 
-export async function getAllEvaluators(): Promise<Evaluator[]> {
+export type EvaluatorDTO = {
+  id: string;
+  createdAt: Date;
+  updatedAt: Date;
+  name: string;
+  type: string;
+  config: LlmConfig | undefined;
+  _count: {
+    evals: number;
+  };
+};
+
+function evaluatorToDTO(dbEvaluator: DbEvaluator): EvaluatorDTO {
+  const config = llmConfigSchema.safeParse(dbEvaluator.config);
+
+  return {
+    ...dbEvaluator,
+    config: config.success ? config.data : undefined, // TODO - throw? report somehow?
+  };
+}
+
+export async function getAllEvaluators(): Promise<EvaluatorDTO[]> {
   await checkRootUser();
 
-  return prisma.evaluator.findMany({
+  const dbEvaluators = await prisma.evaluator.findMany({
     select: evaluatorSelect,
     orderBy: {
       name: "asc",
     },
   });
+
+  return dbEvaluators.map(evaluatorToDTO);
 }
 
-export async function getEvaluatorById(id: string): Promise<Evaluator> {
+export async function getEvaluatorById(id: string): Promise<EvaluatorDTO> {
   await checkRootUser();
 
-  return prisma.evaluator.findUniqueOrThrow({
+  const dbEvaluator = await prisma.evaluator.findUniqueOrThrow({
     where: { id },
     select: evaluatorSelect,
   });
+
+  return evaluatorToDTO(dbEvaluator);
 }
