+++ internal-packages/ai/src/server.ts
@@ -2,8 +2,7 @@
 export { createSquiggleWorkflowTemplate } from "./workflows/createSquiggleWorkflowTemplate.js";
 export { fixSquiggleWorkflowTemplate } from "./workflows/fixSquiggleWorkflowTemplate.js";
 
-export { Workflow } from "./workflows/Workflow.js";
-
+export { llmConfigSchema, Workflow } from "./workflows/Workflow.js";
 export { makeAiCodec } from "./serialization.js";
 
 export { CodeArtifact, PromptArtifact, SourceArtifact } from "./Artifact.js";
